requirements.txt
asgiref==3.7.2
Django==4.2.5
django-admin-sortable2==2.1.9
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
phonenumberslite==8.13.19
Pillow==10.0.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
